CODE: CLI Refactoring
Submission 522596 by nowind_lee to the challenge
http://www.topcoder.com/challenges/30063470
