+++ mythril/ipc/utils.py
@@ -1,3 +1,6 @@
+import os
+import sys
+
 from .constants import BLOCK_TAGS
 
 
@@ -7,6 +10,11 @@ def hex_to_dec(x):
     '''
     return int(x, 16)
 
+def to_bytes(text):
+    return text.encode('utf-8')
+
+def to_text(byte_array):
+    return byte_array.decode('utf-8')
 
 def clean_hex(d):
     '''
@@ -36,3 +44,43 @@ def ether_to_wei(ether):
     Convert ether to wei
     '''
     return ether * 10**18
+
+
+def get_default_ipc_path(testnet=False):
+    if testnet:
+        testnet = "testnet"
+    else:
+        testnet = ""
+
+    if sys.platform == 'darwin':
+        ipc_path = os.path.expanduser(os.path.join("~", "Library", "Ethereum", testnet, "geth.ipc"))
+        if os.path.exists(ipc_path):
+            return ipc_path
+
+        ipc_path = os.path.expanduser(os.path.join("~", "Library", "Application Support", "io.parity.ethereum", "jsonrpc.ipc"))
+        if os.path.exists(ipc_path):
+            return ipc_path
+
+    elif sys.platform.startswith('linux'):
+        ipc_path = os.path.expanduser(os.path.join("~", ".ethereum", testnet, "geth.ipc"))
+        if os.path.exists(ipc_path):
+            return ipc_path
+
+        ipc_path = os.path.expanduser(os.path.join("~", ".local", "share", "io.parity.ethereum", "jsonrpc.ipc"))
+        if os.path.exists(ipc_path):
+            return ipc_path
+
+    elif sys.platform == 'win32':
+        ipc_path = os.path.join("\\\\", ".", "pipe", "geth.ipc")
+        if os.path.exists(ipc_path):
+            return ipc_path
+
+        ipc_path = os.path.join("\\\\", ".", "pipe", "jsonrpc.ipc")
+        if os.path.exists(ipc_path):
+            return ipc_path
+
+    else:
+        raise ValueError(
+            "Unsupported platform '{0}'.  Only darwin/linux2/win32 are "
+            "supported.  You must specify the ipc_path".format(sys.platform)
+        )
